# docs/swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Server API",
    "version": "1.0.0",
    "description": "These are the endpoints for the Connector Server API",
    "contact": {
      "name": "Alessandro Cannarella",
      "email": "cannarella.dev@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/brokers": {
      "get": {
        "tags": [
          "brokers"
        ],
        "summary": "Get brokers",
        "description": "Returns the list of brokers registered in the system",
        "operationId": "getBrokers",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/BrokerDocument"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "brokers"
        ],
        "summary": "Register a broker",
        "description": "Register a broker in the system to be able to publish offers in the federation",
        "operationId": "registerBroker",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Broker name",
            "required": true,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "description": "The broker endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "brokers"
        ],
        "summary": "Unregister a broker",
        "description": "Unregister a broker from the system",
        "operationId": "unregisterBroker",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Broker ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/cluster/init": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "Get ready catalog",
        "operationId": "getReadyCatalog",
        "description": "Get the information for the readiness of the catalog, so if it is configured and ready to be used",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ready": {
                      "type": "boolean"
                    }
                  },
                  "example": {
                    "ready": true
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "Initiate catalog",
        "description": "Initiate the catalog with the information of the cluster: Pretty Name and Contract Endpoint",
        "operationId": "initCatalog",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "prettyname",
            "in": "query",
            "description": "Cluster pretty name",
            "required": true,
            "type": "string"
          },
          {
            "name": "endpoint",
            "in": "query",
            "description": "Contract endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/cluster/prettyname": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "Get cluster pretty name",
        "description": "Get the Cluster pretty name",
        "operationId": "getClusterPrettyName",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prettyName": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "prettyName": "Example Cluster"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "Set cluster pretty name",
        "description": "Set the Cluster pretty name",
        "operationId": "setClusterPrettyName",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Cluster pretty name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/cluster/contractendpoint": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "Get contract endpoint",
        "description": "Get the Cluster contract endpoint",
        "operationId": "getContractEndpoint",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contractEndpoint": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "contractEndpoint": "http://example.com/contracts"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "Set contract endpoint",
        "description": "Set the Cluster contract endpoint",
        "operationId": "setContractEndpoint",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Contract endpoint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BrokerDocument": {
      "type": "object",
      "properties": {
        "brokerID": {
          "type": "string",
          "example": "6d6a3b9e-9f9e-4a0e-8d1a-5a0f0a2e1a0a"
        },
        "brokerName": {
          "type": "string",
          "example": "Example Broker"
        },
        "brokerEndpoint": {
          "type": "string",
          "example": "http://broker.example.com:8080"
        },
        "jwt-token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJib3QiOiJleGFtcGxlLmNvbSIsInNhb"
        },
        "subscribed": {
          "type": "boolean",
          "example": true
        }
      },
      "required": [
        "brokerID",
        "brokerName",
        "brokerEndpoint",
        "jwt-token",
        "subscribed"
      ]
    }
  }
}